backend/notify-service/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Configuracion
pydantic-settings==2.2.1

# Base de datos
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# JWT
python-jose[cryptography]==3.3.0

# Utilidades
python-multipart==0.0.9
